fixed bug to nested rollback
